explore: Simplifies Exploratory Data Analysis

Interactive data exploration with one line of code, automated reporting or use an easy to remember set of tidy functions for low code exploratory data analysis.

See also

Author

Maintainer: Roland Krasser roland.krasser@gmail.com