Writes settings in a table for Explore

`settingsExplore()` defines the parameters used by Explore

Usage

settingsExplore(
  settings,
  output_path,
  file_name,
  train_data = NULL,
  OutputFile = NULL,
  StartRulelength = 1,
  EndRulelength = 3,
  OperatorMethod = "EXHAUSTIVE",
  CutoffMethod = NULL,
  ClassFeature,
  PositiveClass,
  FeatureInclude = "",
  Maximize = "BALANCEDACCURACY",
  Accuracy = 0,
  BalancedAccuracy = 0,
  Specificity = 0,
  OutputMethod = "BEST",
  PrintSettings = "yes",
  PrintPerformance = "no",
  PrintCutoffSets = "no",
  Subsumption = "no",
  BranchBound = "yes",
  Parallel = "no",
  ParallelMethod = "TWO",
  BinaryReduction = "no"
)

Arguments

settings

Settings parameter

output_path

A string declaring the path to the settings

file_name

A string declaring the the path to the file name

train_data

Train data

OutputFile

A string declaring the path to the output file

StartRulelength

Positive integer

EndRulelength

Positive integer

OperatorMethod

One of list with strings, e.g. list = "EXHAUSTIVE", ...

CutoffMethod

One of list with strings, list = "RVAC", ...

ClassFeature

String, should be name of one of columns in data train.

PositiveClass

1 or string (?) (should be one of elements of column 'ClassFeature' in data train)

FeatureInclude

Empty or string (should be name of one of columns in data train)

Maximize

One of list with strings, list = "ACCURACY", ...

Accuracy

Float 0-1 -> default = 0 (if 0, make empty = computationally more beneficial)

Specificity

float 0-1, default = 0

OutputMethod

string EVERY, BEST, INCREMENT

PrintSettings

True or False

PrintPerformance

True or False

Subsumption

True or False

BranchBound

True or False

Parallel

True or False

Value

Settings path