Finds decision rule using Explore

`trainExplore()` finds the best decision rule for the given data set based using EXPLORE.

Usage

trainExplore(
  train_data = NULL,
  settings_path = NULL,
  output_path,
  file_name = "train_data",
  OutputFile = NULL,
  StartRulelength = 1,
  EndRulelength = 3,
  OperatorMethod = "EXHAUSTIVE",
  CutoffMethod = "RVAC",
  ClassFeature = "'class'",
  PositiveClass = "'Iris-versicolor'",
  FeatureInclude = "",
  Maximize = "ACCURACY",
  Accuracy = 0,
  Specificity = 0,
  PrintSettings = TRUE,
  PrintPerformance = TRUE,
  Subsumption = TRUE,
  BranchBound = TRUE,
  Parallel = FALSE
)

Arguments

train_data

Train data

settings_path

A string declaring the path to the settings

output_path

A string declaring the path to the settings

file_name

A string declaring the the path to the file name

OutputFile

A string declaring the path to the output file

StartRulelength

Positive integer

EndRulelength

Positive integer

OperatorMethod

One of list with strings, e.g. list = "EXHAUSTIVE", ...

CutoffMethod

One of list with strings, list = "RVAC", ...

ClassFeature

String, should be name of one of columns in data train. Always provided by the user. The string should be enclused in single quotation marks, e.g. 'class'

PositiveClass

1 or string (?) (should be one of elements of column 'ClassFeature' in data train). Always provided by the user. The string should be enclused in single quotation marks, e.g. 'class'

FeatureInclude

Empty or string (should be name of one of columns in data train)

Maximize

One of list with strings, list = "ACCURACY", ...

Accuracy

Float 0-1 -> default = 0 (if 0, make empty = computationally more beneficial)

Specificity

float 0-1, default = 0

PrintSettings

True or False

PrintPerformance

True or False

Subsumption

True or False

BranchBound

True or False

Parallel

True or False

Value

Model